rename/#30 코드리뷰 반영 - import문 수정
- 수정된 파일명으로 import문 수정
