+++ domain/repositories/ISuccessDayRepository.ts
