+++ domain/repositories/index.ts
@@ -1,2 +1,2 @@
 export * from "@/domain/repositories/IUserRepository";
-export * from "@/domain/repositories/successday-repository";
+export * from "@/domain/repositories/ISuccessDayRepository";
